+++ rusl/src/unistd/mount.rs
@@ -26,26 +26,13 @@ where
             target.exec_with_self_as_ptr(|tgt| {
                 if let Some(data) = data {
                     data.exec_with_self_as_ptr(|data| {
-                        let res = syscall!(
-                            MOUNT,
-                            src,
-                            tgt,
-                            fs_type.label().0.as_ptr(),
-                            flags as usize,
-                            data
-                        );
+                        let res =
+                            syscall!(MOUNT, src, tgt, fs_type.label().0.as_ptr(), flags, data);
                         bail_on_below_zero!(res, "`MOUNT` syscall failed");
                         Ok(res)
                     })
                 } else {
-                    let res = syscall!(
-                        MOUNT,
-                        src,
-                        tgt,
-                        fs_type.label().0.as_ptr(),
-                        flags as usize,
-                        0
-                    );
+                    let res = syscall!(MOUNT, src, tgt, fs_type.label().0.as_ptr(), flags, 0);
                     bail_on_below_zero!(res, "`MOUNT` syscall failed");
                     Ok(res)
                 }
