+++ rusl/src/unistd/swapon.rs
@@ -9,7 +9,7 @@ use sc::syscall;
 pub fn swapon(path: impl AsUnixStr, flags: i32) -> Result<()> {
     path.exec_with_self_as_ptr(|ptr| {
         unsafe {
-            let res = syscall!(SWAPON, ptr, flags as usize);
+            let res = syscall!(SWAPON, ptr, flags);
             bail_on_below_zero!(res, "`SWAPON` syscall failed");
         }
         Ok(())
